feat(revit): improvements to nested family instances + tests
