+++ Objects/Converters/ConverterRevit/ConverterRevitTests/SpeckleConversionTest.cs
@@ -72,27 +72,37 @@ namespace ConverterRevitTests
 
       ConverterRevit converter = new ConverterRevit();
       converter.SetContextDocument(doc);
-      var spkElems = elements.Select(x => converter.ConvertToSpeckle(x)).ToList();
+      //setting context objects for nested routine
+      converter.SetContextObjects(elements.Select(obj => new ApplicationPlaceholderObject { applicationId = obj.UniqueId }).ToList());
+      var spkElems = elements.Select(x => converter.ConvertToSpeckle(x)).Where(x => x != null).ToList();
 
       converter = new ConverterRevit();
       converter.SetContextDocument(fixture.NewDoc);
+      //setting context objects for update routine
       if (appPlaceholders != null)
         converter.SetContextObjects(appPlaceholders);
 
-      //var revitEls = new List<object>();
       var resEls = new List<object>();
+      //used to associate th nested Base objects with eh flat revit ones
+      var flatSpkElems = new List<Base>();
 
       xru.RunInTransaction(() =>
       {
-        //revitEls = spkElems.Select(x => kit.ConvertToNative(x)).ToList();
         foreach (var el in spkElems)
         {
           var res = converter.ConvertToNative(el);
           if (res is List<ApplicationPlaceholderObject> apls)
           {
             resEls.AddRange(apls);
+            flatSpkElems.Add(el);
+            if (el["elements"] != null)
+              flatSpkElems.AddRange(el["elements"] as List<Base>);
           }
-          else resEls.Add(res);
+          else
+          {
+            resEls.Add(res);
+            flatSpkElems.Add(el);
+          }
         }
       }, fixture.NewDoc).Wait();
 
@@ -100,12 +110,8 @@ namespace ConverterRevitTests
 
       for (var i = 0; i < spkElems.Count; i++)
       {
-        var sourceElem = (T)(object)elements[i];
+        var sourceElem = (T)(object)elements.FirstOrDefault(x => x.UniqueId == flatSpkElems[i].applicationId);
         var destElement = (T)((ApplicationPlaceholderObject)resEls[i]).NativeObject;
-        // T destElement;
-        // if (resEls[i] is ApplicationPlaceholderObject apl) destElement = (T)apl.NativeObject;
-        // else destElement = (T)resEls[i];
-
         assert(sourceElem, destElement);
       }
 
@@ -227,7 +233,7 @@ namespace ConverterRevitTests
       {
         var e1 = fixture.SourceDoc.GetElement(expecedParam.AsElementId());
         var e2 = fixture.NewDoc.GetElement(actual.get_Parameter(param).AsElementId());
-        if(e1 is Level l1 && e2 is Level l2)
+        if (e1 is Level l1 && e2 is Level l2)
           Assert.Equal(l1.Elevation, l2.Elevation, 3);
         else if (e1 != null && e2 != null)
           Assert.Equal(e1.Name, e2.Name);
