--- Objects/Converters/ConverterRevit/TestModels/schema-builder.3dm.rhl
@@ -1,3 +1,0 @@
-﻿Admin
-BEAST
-Tuesday, December 8, 2020
